fix(ci): python-coverage-shards — split oversized pre-push validation files under the maintainability line limit
Extract the dirty-finalize implementation (_operation_owned_delta_paths,
_committed_delta_paths, _try_finalize_pre_push_dirty_repair_state,
_rollback_finalize_dirty_residue_before_provider_recovery) from
pre_push_validation.py into a new pre_push_validation_dirty_finalize
sub-module (re-exported by the parent to preserve the public surface and
test monkeypatch semantics, mirroring pre_push_validation_fix_pass.py).
Split test_pr_monitor_pre_push_validation_finalize.py by moving the
provider-recovery propagation and post-commit re-validation regression
tests into a new test_pr_monitor_pre_push_validation_finalize_post_commit
sibling; hoist the local _name_status_z helper into the shared
_pre_push_validation_helpers module so both split files can import it.

Both oversized first-party files now stay under the 1,500-line limit
enforced by test_core_decomposition_maintainability, unblocking the
python-coverage-shards (8) check on PR #615.
